requirements.txt
# Core dependencies
python-dotenv==1.0.0          # For managing API keys and environment variables
requests==2.31.0               # For making HTTP requests to APIs

# Twitter/X API
tweepy==4.14.0                 # Official Twitter API library

# Sentiment Analysis
textblob==0.17.1              # Simple sentiment analysis library
vaderSentiment==3.3.2         # VADER sentiment analyzer (good for social media)

# Data processing
pandas==2.1.4                  # Data manipulation and analysis
numpy==1.26.2                  # Numerical computing

# Database
sqlite3                        # Built into Python (no need to install)

# Web scraping (for SEC EDGAR)
beautifulsoup4==4.12.2        # HTML parsing
lxml==4.9.4                    # XML/HTML parser

# Visualization
matplotlib==3.8.2              # Plotting library
plotly==5.18.0                # Interactive visualizations
dash==2.14.2                  # Web dashboard framework

# Utilities
pytz==2023.3                  # Timezone handling
python-dateutil==2.8.2        # Date utilities

# Rate limiting
ratelimit==2.2.1              # API rate limiting decorator

# Logging
colorlog==6.8.0               # Colored logging output
